Rename itemControl.ts to extensions.ts
