+++ packages/smart-forms-renderer/src/utils/index.ts
@@ -26,7 +26,7 @@ export { initialiseQuestionnaireResponse } from './initialise';
 export { createEmptyQrItem, createEmptyQrGroup, updateQrItemsInGroup } from './qrItem';
 export { mapQItemsIndex, getQrItemsIndex } from './mapItem';
 
-export { isSpecificItemControl, getDecimalPrecision } from './itemControl';
+export { isSpecificItemControl, getDecimalPrecision } from './extensions';
 export { isRepeatItemAndNotCheckbox, isHiddenByEnableWhen } from './qItem';
 export { parseDecimalStringToFloat, parseDecimalStringWithPrecision } from './parseInputs';
 export type { ItemToRepopulate } from './repopulateItems';
